% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggiNEXTPD}
\alias{ggiNEXTPD}
\title{Plots the outcome of \code{iNEXTPD} using the \code{ggplot2} package.}
\usage{
ggiNEXTPD(outcome, plot.type = 1:3)
}
\arguments{
\item{outcome}{the outcome of the function \code{iNEXTPD}}

\item{plot.type}{a positive integer or sequence specifying types of curves. There are three
types of plots: sample-size-based rarefaction and extrapolation curve (\code{$RE.plot.size}, \code{plot.type = 1});
sample coverage curve as a function of sample size (\code{$RE.plot.sizeC}, \code{plot.type = 2}); coverage-based
rarefaction and extrapolation curve (\code{$RE.plot.C}, \code{plot.type = 3}). Default is \code{c(1,2,3)}.}
}
\value{
Returns different plots of the estimated diversity curves based on the \code{ggplot2} package. Three choices of \code{plot.type}:
\itemize{
 \item{\code{$RE.plot.size}: size-based rarefaction and extrapolation curve (\code{plot.type = 1}),
 sampling curve depicting phylogenetic diversity estimates as a function of sample size.} \cr
 \item{\code{$RE.plot.sizeC}: sample completeness curve (\code{plot.type = 2}), the curve of sample coverage as a function of sample size.} \cr
 \item{\code{$RE.plot.C}: coverage-based rarefaction and extrapolation curve (\code{plot.type = 3}), sampling curve depicting phylogenetic diversity estimates as a function of
 sample coverage.}
 }
}
\description{
Function \code{ggiNEXTPD} plots the outcome of \code{iNEXTPD} using the \code{ggplot2} package.
}
\examples{
\donttest{
# Datatype: abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- iNEXTPD(data = data, tree = tree, datatype = "abundance", q = c(0, 1, 2))
ggiNEXTPD(out)

# Datatype: incidence_raw data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out <- iNEXTPD(data = data, nT = nT, datatype = "incidence_raw", tree = tree, q = c(0, 1, 2))
ggiNEXTPD(out)
}
}
