% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{iNEXTPD}
\alias{iNEXTPD}
\title{Interpolation (rarefaction) and extrapolation of Chao et al.’s (2010) phylogenetic diversity and mean phylogenetic diversity (phylogenetic Hill numbers)}
\usage{
iNEXTPD(
  data,
  nT,
  datatype = "abundance",
  tree,
  q = c(0, 1, 2),
  reftime = NULL,
  type = "PD",
  endpoint = NULL,
  knots = 40,
  size = NULL,
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances (for abundance data) or species-by-site incidence raw matrix/data.frame (for incidence data).\cr
Abundance data: a species-by-site matrix/data.frame of species abundances. The row (species) names of
data must match the species names in the phylogenetic tree and thus cannot be missing.\cr
Incidence raw data: species-by-site raw incidence matrix/data.frame. When there are N assemblages
and thus N matrices, users must first merge the N matrices by species identity to obtain a large
merged incidence matrix, where the rows of the matrix refer to all species presented in the merged
data. The row (species) names of data must match the species names in the phylogenetic tree and
thus cannot be missing.}

\item{nT}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the number of sampling units in each assemblage.
If \code{names(nT) = NULL}, then assemblage are automatically named as "assemblage1", "assemblage2",..., etc. Ignored if \code{datatype = "abundance"}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species-by-site raw incidence matrix (\code{datatype = "incidence_raw"}). Default is \code{"abundance"}.}

\item{tree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{q}{a nonnegative value or sequence specifying the diversity order. Default is \code{c(0,1,2)}.}

\item{reftime}{a positive value or sequence specifying the reference times for diversity computation. If \code{NULL},
then \code{reftime} is set to be the tree depth of the phylogenetic tree, which is spanned by all the observed species in
the pooled assemblage. Default is \code{NULL}.}

\item{type}{desired diversity type: \code{type = "PD"} for Chao et al. (2010) phylogenetic diversity
and \code{type = "meanPD"} for mean phylogenetic diversity (phylogenetic Hill number). Default is \code{"PD"}.}

\item{endpoint}{a positive integer specifying the endpoint for the rarefaction and extrapolation range.
If \code{NULL}, then \code{endpoint} = double of the reference sample size in each assemblage. It is ignored if \code{size} is given.}

\item{knots}{a positive integer specifying the number of equally-spaced knots between 1 and the \code{endpoint}. Default is 40.}

\item{size}{a sequence of positive integers specifying the sample sizes for which PD or meanPD estimates will be calculated.
If \code{NULL}, then estimates will be calculated for those sample sizes determined by the specified/default \code{endpoint}
and \code{knots}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals.
Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}
}
\value{
Returns a list containing two tables:
\itemize{
 \item{\code{$size_based}: size-based PD or meanPD estimates along with their confidence intervals
 (if \code{nboot > 0}) and relevant statistics information.} \cr
 \item{\code{$coverage_based}: coverage-based diversity estimates along with confidence intervals
 (if \code{nboot > 0}) and relevant statistics information.}
 }
}
\description{
Function \code{iNEXTPD} computes phylogenetic diversity estimates for rarefied samples and extrapolated samples
along with confidence intervals and related coverage estimates based on Chao et al.’s (2010) phylogenetic
diversity (PD) and mean phylogenetic diversity (meanPD). See Chao et al. (2010, 2015, 2016) and
Hsieh and Chao (2017) for pertinent background and methodologies.
}
\examples{
\donttest{
# Datatype: abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- iNEXTPD(data = data, tree = tree, datatype = "abundance", q = c(0, 1, 2), nboot = 30)
out

# Datatype: incidence_raw data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out <- iNEXTPD(data = data, nT = nT, datatype = "incidence_raw", tree = tree, 
q = c(0, 1, 2))
out
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. \emph{Philosophical Transactions of the Royal Society B.}, 365, 3599-3609. \cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015). Rarefaction and extrapolation of phylogenetic diversity. \emph{Methods in Ecology and Evolution}, 6, 380-388.\cr\cr
Chao, A., Chiu C.-H. and Jost L. (2016). Phylogenetic diversity measures and their decomposition: a framework based on Hill numbers. pp. 141-172 in Pellens R. and Grandcolas P. (eds)
\emph{Biodiversity Conservation and Phylogenetic Systematics: Preserving our Evolutionary Heritage in an Extinction Crisis}, Springer. \cr\cr
Hsieh, T. C. and Chao, A. (2017). Rarefaction and extrapolation: making fair comparison of abundance-sensitive phylogenetic diversity among multiple assemblages. \emph{Systematic Biology}, 66, 100-111.
}
