\name{iNEXTPD2-package}
\alias{iNEXTPD2-package}
\alias{iNEXTPD2}
\docType{package}
\title{
  Interpolation and Extrapolation with Phylogenetic Diversity
}
\description{
  Provides functions for phylogenetic diversity analysis. (1) The iNEXT standardization for phylogenetic diversity: computes and plots rarefaction and extrapolation of phylogenetic diversity based on two types of standardization methods (sample size and sample coverage). (2) The phylogenetic diversity estimation: computes and plots asymptotic phylogenetic diversity profile as a function of diversity order (q-profile) or time (time-profile).
}
\details{
  The four main functions are: (1) \code{iNEXTPD} computes the iNEXT standardization for phylogenetic diversity via rarefaction and extrapolation. (2) \code{estimatePD} computes phylogenetic diversity at specified values of sample coverages. (3) \code{PhdAsy} computes asymptotic phylogenetic diversity profiles. (4) \code{PhdObs} computes empirical or observed phylogenetic profiles.
}
\author{
Anne Chao, Yan-Han Chen and Chun-Huo Chiu

Maintainer: Anne Chao <chao@stat.nthu.edu.tw>
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. \emph{Philosophical Transactions of the Royal Society B.}, 365, 3599-3609. \cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015). Rarefaction and extrapolation of phylogenetic diversity. \emph{Methods in Ecology and Evolution}, 6, 380-388.\cr\cr
Chao, A., Chiu C.-H. and Jost L. (2016). Phylogenetic diversity measures and their decomposition: a framework based on Hill numbers. pp. 141-172 in Pellens R. and Grandcolas P. (eds)
\emph{Biodiversity Conservation and Phylogenetic Systematics: Preserving our Evolutionary Heritage in an Extinction Crisis}, Springer. \cr\cr
Hsieh, T. C. and Chao, A. (2017). Rarefaction and extrapolation: making fair comparison of abundance-sensitive phylogenetic diversity among multiple assemblages. \emph{Systematic Biology}, 66, 100-111.
}
\keyword{ abundance data; extrapolation; incidence data; interpolation; phylogenetic diversity; rarefaction;   }


