% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{PDInfo}
\alias{PDInfo}
\title{Summarizes phylogenetic data information.}
\usage{
PDInfo(data, nT, datatype = "abundance", tree, reftime = NULL)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances (for abundance data) or species-by-site incidence raw matrix/data.frame (for incidence data).
See the function \code{\link{iNEXTPD}} for details.}

\item{nT}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the number of sampling units in each assemblage.
If \code{names(nT) = NULL}, then assemblage are automatically named as "assemblage1", "assemblage2",..., etc. Ignored if \code{datatype = "abundance"}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species-by-site raw incidence matrix (\code{datatype = "incidence_raw"}). Default is \code{"abundance"}.}

\item{tree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{reftime}{a positive value or sequence specifying the reference times for diversity computation. If \code{NULL},
then \code{reftime} is set to be the tree depth of the phylogenetic tree, which is spanned by all the observed species in
the pooled assemblage. Default is \code{NULL}.}
}
\value{
Returns a table of phylogenetic data information, including reference sample size (\code{n}), number of sampling units (\code{nT}),
number of observed species (\code{S.obs}), observed total branch length, i.e., Faith’s PD (\code{PD.obs}), the first two rare species
frequency counts and their branch length sums, at specified/default reference time specified in the argument \code{reftime}.
See Chao et al. (2010, 2015) and Hsieh and Chao (2017) for formulas and interpretations.
}
\description{
Function \code{PDInfo} summarizes phylogenetic data statistics for specified/default reference time.
}
\examples{
\donttest{
# Datatype: abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- PDInfo(data = data, datatype = "abundance", tree = tree)
out

# Datatype: incidence_raw data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out <- PDInfo(data = data, nT = nT, datatype = "incidence_raw", tree = tree)
out
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. \emph{Philosophical Transactions of the Royal Society B.}, 365, 3599-3609. \cr\cr
Chao, A., Chiu, C.-H., Hsieh, T. C., Davis, T., Nipperess, D., and Faith, D. (2015). Rarefaction and extrapolation of phylogenetic diversity. \emph{Methods in Ecology and Evolution}, 6, 380-388.\cr\cr
Hsieh, T. C. and Chao, A. (2017). Rarefaction and extrapolation: making fair comparison of abundance-sensitive phylogenetic diversity among multiple assemblages. \emph{Systematic Biology}, 66, 100-111.
}
