% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{PhdObs}
\alias{PhdObs}
\title{Computes observed phylogenetic diversity and phylogenetic Hill numbers}
\usage{
PhdObs(
  data,
  nT,
  datatype = "abundance",
  tree,
  q = seq(0, 2, by = 0.25),
  reftime = NULL,
  type = "PD",
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances (for abundance data) or species-by-site incidence raw matrix/data.frame (for incidence data).
See the function \code{\link{iNEXTPD}} for details.}

\item{nT}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the number of sampling units in each assemblage.
If \code{names(nT) = NULL}, then assemblage are automatically named as "assemblage1", "assemblage2",..., etc. Ignored if \code{datatype = "abundance"}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
or species-by-site raw incidence matrix (\code{datatype = "incidence_raw"}). Default is \code{"abundance"}.}

\item{tree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{q}{a nonnegative value or sequence specifying the diversity order. Default is \code{seq(0, 2, by = 0.25)}.}

\item{reftime}{a positive value or sequence specifying the reference times for diversity computation. If \code{NULL},
then \code{reftime} is set to be the tree depth of the phylogenetic tree, which is spanned by all the observed species in
the pooled assemblage. Default is \code{NULL}.}

\item{type}{desired diversity type: \code{type = "PD"} for Chao et al. (2010) phylogenetic diversity
and \code{type = "meanPD"} for mean phylogenetic diversity (phylogenetic Hill number). Default is \code{"PD"}.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals.
Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}
}
\value{
Returns a table of empirical (observed) phylogenetic diversity (\code{type = "PD"}) or phylogenetic Hill number (\code{type= "meanPD"})
for specified/default order q and reference time.
}
\description{
\code{PhdObs} computes empirical or observed phylogenetic diversity (PD) and phylogenetic Hill
numbers (meanPD, mean phylogenetic diversity) for specified/default order \code{q} and reference
time specified in the argument \code{reftime}. See Chao et al. (2010) for details of PD and meanPD.
}
\examples{
\donttest{
# Datatype: abundance data
data(data.abu)
data <- data.abu$data
tree <- data.abu$tree
out <- PhdObs(data = data, datatype = "abundance", tree = tree,
q = seq(0, 2, by = 0.25))
out

# Datatype: incidence_raw data
data(data.inc)
data <- data.inc$data
tree <- data.inc$tree
nT <- data.inc$nT
out <- PhdObs(data = data, nT = nT, datatype = "incidence_raw",
tree = tree, q = seq(0, 2, by = 0.25))
out
}
}
\references{
Chao, A., Chiu C.-H. and Jost, L. (2010). Phylogenetic diversity measures based on Hill numbers. \emph{Philosophical Transactions of the Royal Society B.}, 365, 3599-3609. \cr\cr
}
